pyproject.toml
[tool.poetry]
name = "pcdf"
version = "0.1.0"
description = ""
authors = ["Tim S <rndmit@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.8.2"
svix-ksuid = "^0.6.2"
typer = "^0.12.3"
kubemodels = {path = ".local/pkg/kubemodels"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
datamodel-code-generator = "^0.25.8"
mypy = "^1.10.1"
sphinx = "^7.3.7"
myst-parser = "^3.0.1"
sphinxawesome-theme = "^5.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-s -v"

[tool.poe.tasks]
test    = "python -m pytest tests/ --cov=cmptk --cov-report html"
ci-test = "python -m pytest tests/ --cov=cmptk --cov-report term --cov-report xml:coverage.xml --junitxml=report.xml"
